+++ scripts/dice/roller.js
@@ -1,6 +1,6 @@
 import {roll_paranoia} from "./roll.js";
 
-export default class roll_builder extends FormApplication {
+export class roll_builder extends FormApplication {
     constructor(object, options) {
         super(object, options);
         this.dice = {
@@ -12,6 +12,7 @@ export default class roll_builder extends FormApplication {
         this.attr;
         this.skill;
         this.modifier;
+        this.reroll = false;
     }
 
     static get defaultOptions() {
@@ -78,6 +79,10 @@ export default class roll_builder extends FormApplication {
                     attr: this.attr,
                     skill: this.skill,
                     node: formData.node,
+                    dice: this.dice,
+                    actor_id: this.actor_id,
+                    modifier: this.modifier,
+                    reroll: this.reroll,
                 },
             );
         } else {
@@ -95,7 +100,7 @@ export default class roll_builder extends FormApplication {
         ChatMessage.create(chat_options);
     }
 
-    async display_roll_dialog(node, computer, actor_id, attr=null, skill=null, modifier=0) {
+    async display_roll_dialog(node, computer, actor_id, attr = null, skill = null, modifier = 0) {
         this.dice = {
             node: node,
             computer: computer,
@@ -106,4 +111,96 @@ export default class roll_builder extends FormApplication {
         this.modifier = modifier;
         await this.render(true);
     }
+
+    async roll(node, computer_dice, attr, skill, dice, actor_id, modifier, reroll) {
+        let formula;
+        if (node > 0) {
+            formula = `${node}dn`;
+        } else {
+            formula = `${node}dx`;
+        }
+        if (computer_dice > 0) {
+            formula += ` + ${computer_dice}dc`
+        }
+
+        let roll = new roll_paranoia(formula);
+        let chat_data;
+
+        if (attr && skill) {
+            chat_data = await renderTemplate(
+                "systems/paranoia/templates/dice/skill_roll.html",
+                {
+                    roll: await roll.get_roll_data(),
+                    attr: attr,
+                    skill: skill,
+                    node: node,
+                    dice: dice,
+                    actor_id: actor_id,
+                    modifier: modifier,
+                    reroll: reroll,
+                },
+            );
+        } else {
+            chat_data = await roll.render();
+        }
+        let chat_options = {
+            speaker: {
+                actor: actor_id,
+            },
+            content: chat_data,
+            isRoll: true,
+            sound: "sounds/dice.wav",
+        };
+
+        ChatMessage.create(chat_options);
+    }
+}
+
+export async function reroll(...args) {
+    let message = args[0];
+    let element = $(message.message.content);
+    // check if we are the author, as only the original user can reroll
+    if (!message.author.isSelf) {
+        console.log("quitting as I am not the author of this message");
+        ui.notifications.error("You cannot re-roll for another user (good try ;))");
+        return;
+    }
+    // extract values to pass to the roller
+    let node = $('.reroll', element).attr('data-node');
+    let computer_dice = 1;
+    let actor_id = $('.reroll', element).attr('data-actor_id');
+    let attr = $('.reroll', element).attr('data-attr');
+    let skill = $('.reroll', element).attr('data-skill');
+    let wounds_value = $('.reroll', element).attr('data-modifier');
+    let dice = {
+        computer: computer_dice,
+    };
+    if (node > 0) {
+        dice['node'] = node;
+        dice['xode'] = 0;
+    } else {
+        dice['xode'] = node;
+        dice['node'] = 0
+    }
+
+    let actor = game.actors.get(actor_id);
+    if (!actor || actor.system.moxie.value < 1) {
+        console.log("aborting r-eroll due to lack of actor or insufficient moxie");
+        ui.notifications.warn("You must have >= 1 moxie to reoll");
+        return
+    }
+    await actor.update({system: {moxie: {value: actor.system.moxie.value - 1}}});
+
+    // create the roll builder and roll it with the data we extracted
+    let builder = new roll_builder();
+    await builder.roll(
+        node,
+        computer_dice,
+        attr,
+        skill,
+        dice,
+        actor_id,
+        wounds_value,
+        true,
+    );
 }