+++ release-notes.md
@@ -1,5 +1,7 @@
 `1.5.6` - ?
 * Update token names to display full name (`NAME-CLEARANCE-HOME SECTOR-CLONE #`) ([#35](https://github.com/wrycu/paranoia/issues/35))
+* Activating a mutant power sends an audio cue ([#40](https://github.com/wrycu/paranoia/issues/40))
+* Activating a mutant power sends an initial notification to let the GM know more is coming ([#42](https://github.com/wrycu/paranoia/issues/42))
 
 `1.5.5` - 2023-08-24
 * Fix for inability to update NPCs