+++ release-notes.md
@@ -2,6 +2,7 @@
 * Update token names to display full name (`NAME-CLEARANCE-HOME SECTOR-CLONE #`) ([#35](https://github.com/wrycu/paranoia/issues/35))
 * Activating a mutant power sends an audio cue ([#40](https://github.com/wrycu/paranoia/issues/40))
 * Activating a mutant power sends an initial notification to let the GM know more is coming ([#42](https://github.com/wrycu/paranoia/issues/42))
+* Fix for multiple combat managers opening after >1 combat
 
 `1.5.5` - 2023-08-24
 * Fix for inability to update NPCs